ethereum: Remove support for old config
