+++ node/Cargo.toml
@@ -41,6 +41,7 @@ toml = "0.5.7"
 shellexpand = "2.1.0"
 diesel = "1.4.6"
 fail = "0.4"
+http = "0.1.21" # must be compatible with the version rust-web3 uses
 
 [dev-dependencies]
 assert_cli = "0.6"
